Update rfmoggy_hostMode.ino To work with ESP32 D1 mini
Code has been modified and tested to work with ESP32 D1 mini, confirmed through SDR as sending signals. Please consider adding this to the project, if I went about this the wrong way, apologies as this is my first time submitting anything to a GIT project and working with an ESP modules. 

Either way thank you for sharing your code as it has been helpful at furthering my knowledge and spurring my interests in this area, 
I having working code currently in python for raspberry pi used to read and transmit .sub files from flipper zero recordings but it is clunky plan on using this to help me further my efforts in creating a DIY cheap and portable alternative that is freely available as I am not satisfied with currently available projects.

Thank you!!!!